+++ src/Animations/Congregation.hpp
@@ -1,5 +1,7 @@
 #include "BaseAnimation.hpp"
 
+#include <hiimjustin000.more_icons/include/MoreIcons.hpp>
+
 struct ObjectProperties {
     int p1 = 1; // id
     float p2 = 0.f; // positionX
@@ -640,6 +642,9 @@ private:
 
         player->updateGlowColor();
         player->updatePlayerGlow();
+        
+        if (Loader::get()->isModLoaded("hiimjustin000.more_icons"))
+            MoreIcons::updatePlayerObject(player);
 
         return player;
     }