+++ changelog.md
@@ -1,3 +1,9 @@
+# v1.0.1
+
+* Fix terraria preview crash
+* Fix terraria invisible player
+* Add more icons support in preview and congregation
+
 # v1.0.0
 
 * firt